extract ensureEnoughSize func
