--- rlp/decode.go
@@ -1,1 +1,0 @@
-package rlp
