+++ txpool/types.go
@@ -24,6 +24,7 @@ import (
 
 	"github.com/holiman/uint256"
 	"github.com/ledgerwatch/erigon-lib/gointerfaces/types"
+	"github.com/ledgerwatch/erigon-lib/rlp"
 	"github.com/ledgerwatch/secp256k1"
 	"golang.org/x/crypto/sha3"
 )
@@ -35,6 +36,7 @@ type Hashes []byte // flatten list of 32-byte hashes
 // TxContext is object that is required to parse transactions and turn transaction payload into TxSlot objects
 // usage of TxContext helps avoid extra memory allocations
 type TxParseContext struct {
+	recCtx        *secp256k1.Context // Context for sender recovery
 	keccak1       hash.Hash
 	keccak2       hash.Hash
 	chainId, r, s uint256.Int // Signature values
@@ -42,7 +44,6 @@ type TxParseContext struct {
 	buf           [65]byte // buffer needs to be enough for hashes (32 bytes) and for public key (65 bytes)
 	sighash       [32]byte
 	sig           [65]byte
-	recCtx        *secp256k1.Context // Context for sender recovery
 }
 
 func NewTxParseContext() *TxParseContext {
@@ -67,6 +68,8 @@ type TxSlot struct {
 	feeCap      uint64      // Maximum fee that transaction burns and gives to the miner/block proposer
 	gas         uint64      // Gas limit of the transaction
 	value       uint256.Int // Value transferred by the transaction
+	idHash      [32]byte    // Transaction hash for the purposes of using it as a transaction Id
+	sender      [20]byte    // Sender address for the transaction, recovered from the signature
 	creation    bool        // Set to true if "To" field of the transation is not set
 	dataLen     int         // Length of transaction's data (for calculation of intrinsic gas)
 	alAddrCount int         // Number of addresses in the access list
@@ -74,101 +77,6 @@ type TxSlot struct {
 	//bestIdx     int         // Index of the transaction in the best priority queue (of whatever pool it currently belongs to)
 	//worstIdx    int         // Index of the transaction in the worst priority queue (of whatever pook it currently belongs to)
 	//local       bool        // Whether transaction has been injected locally (and hence needs priority when mining or proposing a block)
-	idHash [32]byte // Transaction hash for the purposes of using it as a transaction Id
-	sender [20]byte // Sender address for the transaction, recovered from the signature
-}
-
-// beInt parses Big Endian representation of an integer from given payload at given position
-func beInt(payload []byte, pos, length int) (int, error) {
-	var r int
-	if length > 0 && payload[pos] == 0 {
-		return 0, fmt.Errorf("integer encoding for RLP must not have leading zeros: %x", payload[pos:pos+length])
-	}
-	for _, b := range payload[pos : pos+length] {
-		r = (r << 8) | int(b)
-	}
-	return r, nil
-}
-
-// prefix parses RLP prefix from given payload at given position. It returns the offset and length of the RLP element
-// as well as the indication of whether it is a list of string
-func prefix(payload []byte, pos int) (dataPos int, dataLen int, list bool, err error) {
-	switch first := payload[pos]; {
-	case first < 128:
-		dataPos = pos
-		dataLen = 1
-		list = false
-	case first < 184:
-		// string of len < 56, and it is non-legacy transaction
-		dataPos = pos + 1
-		dataLen = int(first) - 128
-		list = false
-	case first < 192:
-		// string of len >= 56, and it is non-legacy transaction
-		beLen := int(first) - 183
-		dataPos = pos + 1 + beLen
-		dataLen, err = beInt(payload, pos+1, beLen)
-		list = false
-	case first < 248:
-		// list of len < 56, and it is a legacy transaction
-		dataPos = pos + 1
-		dataLen = int(first) - 192
-		list = true
-	default:
-		// list of len >= 56, and it is a legacy transaction
-		beLen := int(first) - 247
-		dataPos = pos + 1 + beLen
-		dataLen, err = beInt(payload, pos+1, beLen)
-		list = true
-	}
-	return
-}
-
-// parseUint64 parses uint64 number from given payload at given position
-func parseUint64(payload []byte, pos int) (int, uint64, error) {
-	dataPos, dataLen, list, err := prefix(payload, pos)
-	if err != nil {
-		return 0, 0, err
-	}
-	if list {
-		return 0, 0, fmt.Errorf("uint64 must be a string, not list")
-	}
-	if dataPos+dataLen >= len(payload) {
-		return 0, 0, fmt.Errorf("unexpected end of payload")
-	}
-	if dataLen > 8 {
-		return 0, 0, fmt.Errorf("uint64 must not be more than 8 bytes long, got %d", dataLen)
-	}
-	if dataLen > 0 && payload[dataPos] == 0 {
-		return 0, 0, fmt.Errorf("integer encoding for RLP must not have leading zeros: %x", payload[dataPos:dataPos+dataLen])
-	}
-	var r uint64
-	for _, b := range payload[dataPos : dataPos+dataLen] {
-		r = (r << 8) | uint64(b)
-	}
-	return dataPos + dataLen, r, nil
-}
-
-// parseUint256 parses uint256 number from given payload at given position
-func parseUint256(payload []byte, pos int, x *uint256.Int) (int, error) {
-	dataPos, dataLen, list, err := prefix(payload, pos)
-	if err != nil {
-		return 0, err
-	}
-	if list {
-		return 0, fmt.Errorf("uint256 must be a string, not list")
-	}
-	if dataPos+dataLen > len(payload) {
-		return 0, fmt.Errorf("unexpected end of payload")
-	}
-	if dataLen > 32 {
-		return 0, fmt.Errorf("uint256 must not be more than 8 bytes long, got %d", dataLen)
-	}
-	if dataLen > 0 && payload[dataPos] == 0 {
-		return 0, fmt.Errorf("integer encoding for RLP must not have leading zeros: %x", payload[dataPos:dataPos+dataLen])
-	}
-	x.SetBytes(payload[dataPos : dataPos+dataLen])
-	return dataPos + dataLen, nil
 }
 
 const (
@@ -189,11 +97,11 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 	ctx.keccak1.Reset()
 	ctx.keccak2.Reset()
 	var slot TxSlot
-	// Legacy transations have list prefix, whereas EIP-2718 transactions have string prefix
-	// therefore we assign the first returned value of prefix function (list) to legacy variable
-	dataPos, dataLen, legacy, err := prefix(payload, pos)
+	// Legacy transations have list ParsePrefix, whereas EIP-2718 transactions have string ParsePrefix
+	// therefore we assign the first returned value of ParsePrefix function (list) to legacy variable
+	dataPos, dataLen, legacy, err := rlp.ParsePrefix(payload, pos)
 	if err != nil {
-		return nil, 0, fmt.Errorf("%s: size prefix: %v", ParseTransactionErrorPrefix, err)
+		return nil, 0, fmt.Errorf("%s: size ParsePrefix: %v", ParseTransactionErrorPrefix, err)
 	}
 	payloadLen := len(payload)
 	if dataPos+dataLen != payloadLen {
@@ -207,18 +115,18 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 	if !legacy {
 		txType = int(payload[p])
 		if _, err = ctx.keccak1.Write(payload[p : p+1]); err != nil {
-			return nil, 0, fmt.Errorf("%s: computing idHash (hashing type prefix): %w", ParseTransactionErrorPrefix, err)
+			return nil, 0, fmt.Errorf("%s: computing idHash (hashing type ParsePrefix): %w", ParseTransactionErrorPrefix, err)
 		}
 		if _, err = ctx.keccak2.Write(payload[p : p+1]); err != nil {
-			return nil, 0, fmt.Errorf("%s: computing signHash (hashing type prefix): %w", ParseTransactionErrorPrefix, err)
+			return nil, 0, fmt.Errorf("%s: computing signHash (hashing type ParsePrefix): %w", ParseTransactionErrorPrefix, err)
 		}
 		p++
 		if p >= payloadLen {
 			return nil, 0, fmt.Errorf("%s: unexpected end of payload after txType", ParseTransactionErrorPrefix)
 		}
-		dataPos, dataLen, list, err = prefix(payload, p)
+		dataPos, dataLen, list, err = rlp.ParsePrefix(payload, p)
 		if err != nil {
-			return nil, 0, fmt.Errorf("%s: envelope prefix: %v", ParseTransactionErrorPrefix, err)
+			return nil, 0, fmt.Errorf("%s: envelope ParsePrefix: %v", ParseTransactionErrorPrefix, err)
 		}
 		if !list {
 			return nil, 0, fmt.Errorf("%s: envelope must be a list, not string", ParseTransactionErrorPrefix)
@@ -236,7 +144,7 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 	sigHashPos := p
 	// If it is non-legacy tx, chainId follows, but we skip it
 	if !legacy {
-		dataPos, dataLen, list, err = prefix(payload, p)
+		dataPos, dataLen, list, err = rlp.ParsePrefix(payload, p)
 		if err != nil {
 			return nil, 0, fmt.Errorf("%s: chainId len: %w", ParseTransactionErrorPrefix, err)
 		}
@@ -249,13 +157,13 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 		p = dataPos + dataLen
 	}
 	// Next follows the nonce, which we need to parse
-	p, slot.nonce, err = parseUint64(payload, p)
+	p, slot.nonce, err = rlp.ParseUint64(payload, p)
 	if err != nil {
 		return nil, 0, fmt.Errorf("%s: nonce: %w", ParseTransactionErrorPrefix, err)
 	}
 	// Next follows gas price or tip
 	// Although consensus rules specify that tip can be up to 256 bit long, we narrow it to 64 bit
-	p, slot.tip, err = parseUint64(payload, p)
+	p, slot.tip, err = rlp.ParseUint64(payload, p)
 	if err != nil {
 		return nil, 0, fmt.Errorf("%s: tip: %w", ParseTransactionErrorPrefix, err)
 	}
@@ -265,18 +173,18 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 		slot.feeCap = slot.tip
 	} else {
 		// Although consensus rules specify that feeCap can be up to 256 bit long, we narrow it to 64 bit
-		p, slot.feeCap, err = parseUint64(payload, p)
+		p, slot.feeCap, err = rlp.ParseUint64(payload, p)
 		if err != nil {
 			return nil, 0, fmt.Errorf("%s: feeCap: %w", ParseTransactionErrorPrefix, err)
 		}
 	}
 	// Next follows gas
-	p, slot.gas, err = parseUint64(payload, p)
+	p, slot.gas, err = rlp.ParseUint64(payload, p)
 	if err != nil {
 		return nil, 0, fmt.Errorf("%s: gas: %w", ParseTransactionErrorPrefix, err)
 	}
 	// Next follows the destrination address (if present)
-	dataPos, dataLen, list, err = prefix(payload, p)
+	dataPos, dataLen, list, err = rlp.ParsePrefix(payload, p)
 	if err != nil {
 		return nil, 0, fmt.Errorf("%s: to len: %w", ParseTransactionErrorPrefix, err)
 	}
@@ -293,12 +201,12 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 	slot.creation = dataLen == 0
 	p = dataPos + dataLen
 	// Next follows value
-	p, err = parseUint256(payload, p, &slot.value)
+	p, err = rlp.ParseUint256(payload, p, &slot.value)
 	if err != nil {
 		return nil, 0, fmt.Errorf("%s: value: %w", ParseTransactionErrorPrefix, err)
 	}
 	// Next goes data, but we are only interesting in its length
-	dataPos, dataLen, list, err = prefix(payload, p)
+	dataPos, dataLen, list, err = rlp.ParsePrefix(payload, p)
 	if err != nil {
 		return nil, 0, fmt.Errorf("%s: data len: %w", ParseTransactionErrorPrefix, err)
 	}
@@ -312,7 +220,7 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 	p = dataPos + dataLen
 	// Next follows access list for non-legacy transactions, we are only interesting in number of addresses and storage keys
 	if !legacy {
-		dataPos, dataLen, list, err = prefix(payload, p)
+		dataPos, dataLen, list, err = rlp.ParsePrefix(payload, p)
 		if err != nil {
 			return nil, 0, fmt.Errorf("%s: access list len: %w", ParseTransactionErrorPrefix, err)
 		}
@@ -325,7 +233,7 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 		tuplePos := dataPos
 		var tupleLen int
 		for tuplePos < dataPos+dataLen {
-			tuplePos, tupleLen, list, err = prefix(payload, tuplePos)
+			tuplePos, tupleLen, list, err = rlp.ParsePrefix(payload, tuplePos)
 			if err != nil {
 				return nil, 0, fmt.Errorf("%s: tuple len: %w", ParseTransactionErrorPrefix, err)
 			}
@@ -336,7 +244,7 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 				return nil, 0, fmt.Errorf("%s: unexpected end of access list after tuple", ParseTransactionErrorPrefix)
 			}
 			var addrPos, addrLen int
-			addrPos, addrLen, list, err = prefix(payload, tuplePos)
+			addrPos, addrLen, list, err = rlp.ParsePrefix(payload, tuplePos)
 			if err != nil {
 				return nil, 0, fmt.Errorf("%s: tuple addr len: %w", ParseTransactionErrorPrefix, err)
 			}
@@ -351,7 +259,7 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 			}
 			slot.alAddrCount++
 			var storagePos, storageLen int
-			storagePos, storageLen, list, err = prefix(payload, addrPos+addrLen)
+			storagePos, storageLen, list, err = rlp.ParsePrefix(payload, addrPos+addrLen)
 			if err != nil {
 				return nil, 0, fmt.Errorf("%s: storage key list len: %w", ParseTransactionErrorPrefix, err)
 			}
@@ -364,7 +272,7 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 			skeyPos := storagePos
 			var skeyLen int
 			for skeyPos < storagePos+storageLen {
-				skeyPos, skeyLen, list, err = prefix(payload, skeyPos)
+				skeyPos, skeyLen, list, err = rlp.ParsePrefix(payload, skeyPos)
 				if err != nil {
 					return nil, 0, fmt.Errorf("%s: tuple storage key len: %w", ParseTransactionErrorPrefix, err)
 				}
@@ -397,7 +305,7 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 	sigHashLen := uint(sigHashEnd - sigHashPos)
 	var chainIdBits, chainIdLen int
 	if legacy {
-		p, err = parseUint256(payload, p, &ctx.chainId)
+		p, err = rlp.ParseUint256(payload, p, &ctx.chainId)
 		if err != nil {
 			return nil, 0, fmt.Errorf("%s: V: %w", ParseTransactionErrorPrefix, err)
 		}
@@ -414,14 +322,14 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 				chainIdLen = 1
 			} else {
 				chainIdLen = (chainIdBits + 7) / 8 // It is always < 56 bytes
-				sigHashLen++                       // For chainId len prefix
+				sigHashLen++                       // For chainId len ParsePrefix
 			}
 			sigHashLen += uint(chainIdLen) // For chainId
 			sigHashLen += 2                // For two extra zeros
 		}
 	} else {
 		var v uint64
-		p, v, err = parseUint64(payload, p)
+		p, v, err = rlp.ParseUint64(payload, p)
 		if err != nil {
 			return nil, 0, fmt.Errorf("%s: V: %w", ParseTransactionErrorPrefix, err)
 		}
@@ -431,12 +339,12 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 		vByte = byte(v)
 	}
 	// Next follows R of the signature
-	p, err = parseUint256(payload, p, &ctx.r)
+	p, err = rlp.ParseUint256(payload, p, &ctx.r)
 	if err != nil {
 		return nil, 0, fmt.Errorf("%s: R: %w", ParseTransactionErrorPrefix, err)
 	}
 	// New follows S of the signature
-	p, err = parseUint256(payload, p, &ctx.s)
+	p, err = rlp.ParseUint256(payload, p, &ctx.s)
 	if err != nil {
 		return nil, 0, fmt.Errorf("%s: S: %w", ParseTransactionErrorPrefix, err)
 	}
@@ -447,19 +355,20 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 		}
 	}
 	ctx.keccak1.Sum(slot.idHash[:0])
+	//ctx.keccak1.(io.Reader).Read(slot.idHash[:32])
 	// Computing sigHash (hash used to recover sender from the signature)
-	// Write len prefix to the sighash
+	// Write len ParsePrefix to the sighash
 	if sigHashLen < 56 {
 		ctx.buf[0] = byte(sigHashLen) + 192
 		if _, err := ctx.keccak2.Write(ctx.buf[:1]); err != nil {
-			return nil, 0, fmt.Errorf("%s: computing signHash (hashing len prefix): %w", ParseTransactionErrorPrefix, err)
+			return nil, 0, fmt.Errorf("%s: computing signHash (hashing len ParsePrefix): %w", ParseTransactionErrorPrefix, err)
 		}
 	} else {
 		beLen := (bits.Len(uint(sigHashLen)) + 7) / 8
 		binary.BigEndian.PutUint64(ctx.buf[1:], uint64(sigHashLen))
 		ctx.buf[8-beLen] = byte(beLen) + 247
 		if _, err := ctx.keccak2.Write(ctx.buf[8-beLen : 9]); err != nil {
-			return nil, 0, fmt.Errorf("%s: computing signHash (hashing len prefix): %w", ParseTransactionErrorPrefix, err)
+			return nil, 0, fmt.Errorf("%s: computing signHash (hashing len ParsePrefix): %w", ParseTransactionErrorPrefix, err)
 		}
 	}
 	if _, err = ctx.keccak2.Write(payload[sigHashPos:sigHashEnd]); err != nil {
@@ -491,6 +400,7 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 		}
 	}
 	// Squeeze sighash
+	//ctx.keccak2.(io.Reader).Read(ctx.sighash[:32])
 	ctx.keccak2.Sum(ctx.sighash[:0])
 	binary.BigEndian.PutUint64(ctx.sig[0:8], ctx.r[3])
 	binary.BigEndian.PutUint64(ctx.sig[8:16], ctx.r[2])
@@ -505,14 +415,15 @@ func (ctx *TxParseContext) ParseTransaction(payload []byte, pos int) (*TxSlot, i
 	if _, err = secp256k1.RecoverPubkeyWithContext(ctx.recCtx, ctx.sighash[:], ctx.sig[:], ctx.buf[:0]); err != nil {
 		return nil, 0, fmt.Errorf("%s: recovering sender from signature: %w", ParseTransactionErrorPrefix, err)
 	}
-	// apply keccak to the public key
+	//apply keccak to the public key
 	ctx.keccak2.Reset()
 	if _, err = ctx.keccak2.Write(ctx.buf[1:65]); err != nil {
 		return nil, 0, fmt.Errorf("%s: computing sender from public key: %w", ParseTransactionErrorPrefix, err)
 	}
 	// squeeze the hash of the public key
 	ctx.keccak2.Sum(ctx.buf[:0])
-	// take last 20 bytes as address
+	//ctx.keccak2.(io.Reader).Read(ctx.buf[:32])
+	//take last 20 bytes as address
 	copy(slot.sender[:], ctx.buf[12:32])
 	return &slot, p, nil
 }
